Fixed various bugs
Fixed a bug that would update the titles after a second search, but would still add titles from the previous search to corpus. Did the similar for the Clear Results button. Fixed a problem with myBasket where the Remove function would remove a movie from the list, but not from the corpus itself. Updated some tooltips.